separate test for token-awareness
